Rearranged quite some files in order to have meaningful names
Backup and Restore jobs are now templates, so you can 'add them' to the
project at will, and they will do just that, backup or restore the data.

There is also an extra deploymentconfig that will bring up a crond
scheduler that can be used to run scripts with the oc binary. All
scripts ran there have the full container environment.

Tite